pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reviewer"
version = "1.0.0"
description = "Deep Research Report Reviewer — LangGraph subgraph for structured feedback"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "pydantic>=2.0",
    "python-dotenv",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: tests that call real LLM API (deselect with '-m not integration')",
]
addopts = "-m 'not integration'"
